+++ tests/php/bootstrap.php
@@ -100,10 +100,6 @@ namespace Magento\Framework\App {
         }
     }
 
-    interface RequestInterface
-    {
-        public function getParam($key, $defaultValue = null);
-    }
 }
 
 namespace Magento\Framework\App\Request {
@@ -200,13 +196,6 @@ namespace Magento\Framework\Data\Collection {
     }
 }
 
-namespace Magento\Framework\Data {
-    interface OptionSourceInterface
-    {
-        public function toOptionArray(): array;
-    }
-}
-
 namespace Magento\Framework\Exception {
     class LocalizedException extends \Exception
     {
@@ -236,7 +225,6 @@ namespace Magento\Framework\View\Element\Block {
 
 namespace {
     use Magento\Framework\App\Config\ScopeConfigInterface;
-    use Magento\Framework\App\RequestInterface;
     use Magento\Store\Model\ScopeInterface;
 
     if (!function_exists('__')) {
@@ -324,19 +312,6 @@ namespace {
         public function getStore($code = null): BasicrumTestStore
         {
             return new BasicrumTestStore($this->storeWebsites[(string) $code] ?? 'base');
-        }
-    }
-
-    final class BasicrumTestRequest implements RequestInterface
-    {
-        /** @param array<string, mixed> $params */
-        public function __construct(private array $params = [])
-        {
-        }
-
-        public function getParam($key, $defaultValue = null)
-        {
-            return $this->params[$key] ?? $defaultValue;
         }
     }